fixes to task3
move components to separate folders
move css to separate files
move logic to functions in utils -> fetching data
apply eslint rules
key:value object -> release_date
remove console.log()
MovieTile.jsx as functional component
MovieDetails.jsx calculate on beginig of render
SearchForm.jsx >> handleFormSubmit no preventDefault
setState as functional call
destructuring of props
