+++ src/components/Body/MovieGenre/MovieGenre.Styles.jsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const MovieGenreWrapper = styled.p`
+  font-size: 10px;
+  color: black;
+  margin: 0px;
+`;
+
+export default MovieGenreWrapper;
