+++ src/components/Body/MovieInfo/MovieInfo.Styles.jsx
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+
+export const MovieInfoWrapper = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+`;
+
+export const Title = styled.p`
+  font-size: 14px;
+  color: black;
+  margin: 5px;
+`;
+
+export const Year = styled.p`
+  color: black;
+  font-size: 14px;
+  margin: 5px;
+`;
