+++ src/components/Body/MoviePoster/MoviePoster.Styles.jsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const PosterWrapper = styled.div`
+  width: 200px;
+  height: 300px;
+  overflow: hidden;
+`;
+
+export const Img = styled.img`
+  width: 100%;
+`;
