+++ src/components/Body/MoviePoster/MoviePoster.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { PosterWrapper, Img } from './MoviePoster.Styles';
+
+const MoviePoster = ({ posterImage }) => (
+  <PosterWrapper>
+    <Img src={posterImage} />
+  </PosterWrapper>
+);
+
+export default MoviePoster;
