+++ src/components/Body/MovieTile/MovieTile.Styles.jsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const MovieTileWrapper = styled.div`
+  width: 200px;
+  height: 380px;
+  display: flex;
+  flex-direction: column;
+  margin-top: 10px;
+  background-color: #fff;
+  border-radius: 6px;
+`;
+
+export default MovieTileWrapper;
