+++ src/components/Body/Results/Results.Styles.jsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const ResultsWrapper = styled.div`
+  grid-area: 3 / 2 / 4 / 3;
+  min-height: 500px;
+  display: flex;
+  flex-wrap: wrap;
+  align-content: flex-start;
+  justify-content: space-around;
+  align-content: flex-start;
+`;
+
+export default ResultsWrapper;
