+++ src/components/Detail/MovieDetails.Styles.jsx
@@ -1,0 +1,32 @@
+import styled from 'styled-components';
+
+export const DetailsWrapper = styled.div`
+  width: 800px;
+  height: 400px;
+  display: flex;
+`;
+
+export const InfoWrapper = styled.div`
+  width: 600px;
+  margin: 16px;
+`;
+
+export const MovieTitle = styled.p`
+  font-size: 32px;
+  color: #dd1133;
+`;
+
+export const MovieRating = styled.p`
+  font-size: 18px;
+  color: white;
+`;
+
+export const MovieData = styled.p`
+  font-size: 14px;
+  color: white;
+`;
+
+export const MovieSummary = styled.p`
+  font-size: 14px;
+  color: white;
+`;
