+++ src/components/Header/FormTitle/FormTitle.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import H2 from './FormTitle.Styles';
+
+const FormTitle = () => <H2>FIND YOUR MOVIE</H2>;
+
+export default FormTitle;
