+++ src/components/Header/PageName/PageName.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import H1 from './PageName.Styles';
+
+const PageName = () => <H1>Netflix Roulette</H1>;
+
+export default PageName;
