+++ src/components/Helper/ResultsOptions/ResultsOptions.Styles.jsx
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+export const ResultsOptionsWrapper = styled.div`
+  grid-area: 2 / 2 / 3 / 3;
+  display: flex;
+  justify-content: space-between;
+  background-color: #eee;
+`;
+
+export const ButtonsWrapper = styled.div`
+  display: flex;
+`;
+
+export const Display = styled.p`
+  font-size: 18px;
+  margin: 5px;
+  padding: 5px;
+`;
