+++ src/utils/sortingTypeConstants.js
@@ -1,0 +1,9 @@
+export const sortingTypeForSearch = {
+  'release date': 'release_date',
+  rating: 'rating'
+};
+
+export const sortingTypeForDisplay = {
+  'release date': 'release date',
+  rating: 'rating'
+};
